@@ -7,8 +7,8 @@
     "format": "prettier --write \"**/*.{js,json,md,html,css,scss,ts}\"",
     "prepare": "husky install",
     "update:git": "git add -A && git commit -m 'update' && git push",
-    "files": "node server/writeAllFilePaths.js && prettier src/files.ts --write && git add src/files.ts",
-    "start": "webpack --env=development",
+    "files": "node server/writeAllFilePaths.js && prettier source/components/files.ts --write",
+    "start": "webpack serve --env=development",
     "build": "webpack --env=production",
     "prepare:commit": "npm run files",
     "netlify": "npm ci && npm run build && rm -r node_modules && cd threshold && npm run netlify:website"
@@ -17,46 +17,49 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
+    "animate.css": "^4.1.1",
     "buffer": "^6.0.3",
     "dataframe-js": "^1.4.4",
-    "jszip": "^3.7.1",
-    "papaparse": "^5.3.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-dropzone": "^12.0.4",
-    "sweetalert2": "^11.4.4"
+    "jszip": "^3.9.1",
+    "papaparse": "^5.3.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-dropzone": "^14.2.1",
+    "sweetalert2": "^11.4.10"
   },
   "devDependencies": {
-    "@babel/core": "^7.17.5",
-    "@babel/preset-env": "^7.16.11",
+    "@babel/core": "^7.17.10",
+    "@babel/preset-env": "^7.17.10",
     "@babel/preset-react": "^7.16.7",
-    "@types/bootstrap": "^5.1.9",
+    "@types/bootstrap": "^5.1.11",
     "@types/dropzone": "^5.7.4",
     "@types/papaparse": "^5.3.2",
-    "@types/react": "^17.0.40",
-    "@types/react-dom": "^17.0.13",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.3",
     "@types/webfontloader": "^1.6.34",
-    "@typescript-eslint/eslint-plugin": "^5.14.0",
-    "@typescript-eslint/parser": "^5.14.0",
-    "babel-loader": "^8.2.3",
+    "@typescript-eslint/eslint-plugin": "^5.22.0",
+    "@typescript-eslint/parser": "^5.22.0",
+    "babel-loader": "^8.2.5",
     "clean-webpack-plugin": "^4.0.0",
     "css-loader": "^6.7.1",
-    "eslint": "^8.11.0",
-    "eslint-plugin-react": "^7.29.3",
-    "express": "^4.17.3",
+    "eslint": "^8.15.0",
+    "eslint-plugin-react": "^7.29.4",
+    "express": "^4.18.1",
     "husky": "^7.0.4",
-    "lint-staged": "^12.3.5",
+    "lint-staged": "^12.4.1",
     "mini-css-extract-plugin": "^2.6.0",
     "postcss-loader": "^6.2.1",
-    "prettier": "^2.5.1",
-    "sass": "^1.49.9",
+    "prettier": "^2.6.2",
+    "sass": "^1.51.0",
     "sass-loader": "^12.6.0",
     "style-loader": "^3.3.1",
-    "ts-loader": "^9.2.8",
-    "typescript": "^4.6.2",
-    "webpack": "^5.70.0",
+    "ts-loader": "^9.3.0",
+    "typescript": "^4.6.4",
+    "webpack": "^5.72.0",
     "webpack-cli": "^4.9.2",
-    "xlsx": "^0.18.3"
+    "webpack-dev-middleware": "^5.3.1",
+    "webpack-dev-server": "^4.9.0",
+    "xlsx": "^0.18.5"
   },
   "lint-staged": {
     "*.{js,json,md,html,css,scss,ts}": "prettier --write"
