@@ -2,6 +2,10 @@ import React, { Component } from "react";
 
 export default class Running extends Component {
   render() {
-    return <div>Running</div>;
+    return (
+      <>
+        <p className="emphasize">Upload successful!</p>
+      </>
+    );
   }
 }
